AgentProcessor: three-component composition (turn loop, LLM request, codemode)
The processor class is now ~100 lines of wiring over three components that
share the pure fold and a small host surface: AgentTurnLoop (mirror,
waiting-clear, interrupt, error transcription, at-head lifecycle),
AgentLlmRequest (transport attempts, chunk streaming, settle appends, and
compaction), and AgentCodemode (slash commands, response parsing through the
injected AgentResponseFormat, settlement rendering + spill). Idempotency keys
mint through the host in the fixed agent/ namespace so a stream handed
between the classic processor and a future variant dedupes consequences
instead of re-executing scripts. Zero behavior change; the existing suites
pass unmodified.
